refactor(Diaries) : Diaries 엔티티 JSON 형태로 저장하기 위한 애노테이션 추가
